+++ eventstore/acceptanece_testing.go
@@ -70,7 +70,7 @@ func AcceptanceTest(t *testing.T, ctx context.Context, store eh.EventStore) []eh
 	// Save event, version 2, with metadata.
 	event2 := eh.NewEvent(mocks.EventType, &mocks.EventData{Content: "event2"}, timestamp,
 		eh.ForAggregate(mocks.AggregateType, id, 2),
-		eh.WithMetadata(map[string]interface{}{"meta": "data", "num": int32(42)}),
+		eh.WithMetadata(map[string]interface{}{"meta": "data", "num": 42.0}),
 	)
 	err = store.Save(ctx, []eh.Event{event2}, 1)
 	if err != nil {
